% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuranker_auth.R
\name{accuranker_auth}
\alias{accuranker_auth}
\title{Authorize Accuranker API with your token}
\usage{
accuranker_auth(
  client_id = Sys.getenv("ACCURANKER_CLIENT_ID"),
  client_secret = Sys.getenv("ACCURANKER_CLIENT_SECRET"),
  token = NULL,
  new_user = FALSE,
  silent = FALSE
)
}
\description{
Enter your token that can be found on https://app.accuranker.com/app/profile.
}
\examples{
accuranker_auth(client_id = Sys.getenv("ACCURANKER_CLIENT_ID"), client_secret = Sys.getenv("ACCURANKER_CLIENT_SECRET"), token = NULL, new_user = FALSE, silent = FALSE)
}
