% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuranker_competitor_ranks.R
\name{accuranker_competitor_ranks}
\alias{accuranker_competitor_ranks}
\title{Get the keyword rankings of your competitors for a specific domain}
\usage{
accuranker_competitor_ranks(
  domain_name,
  date_range = "TODAY",
  silent = FALSE,
  simple_token = NULL
)
}
\arguments{
\item{date_range}{The time period for which to get rankings from}

\item{silent}{If TRUE, nothing will be printed to the console}

\item{simple_token}{Can be used to insert a custom API token instead of using oauth2.}

\item{domain_id}{A specific domain id}
}
\description{
Get the keyword rankings of your competitors for a specific domain
}
\examples{
accuranker_keywords(domain_name, date_range = "TODAY", silent = FALSE, simple_token = NULL)
}
