% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuranker_custom_api.R
\name{accuranker_custom_api}
\alias{accuranker_custom_api}
\title{Collect custom data from the Accuranker API}
\usage{
accuranker_custom_api(request, simple_token = NULL)
}
\arguments{
\item{request}{A string that contains the custom API request for Accuranker.}

\item{simple_token}{Can be used to insert a custom API token instead of using oauth2.}
}
\description{
This function will import data for the chosen parameters.
}
\examples{
accuranker_custom_api(request, simple_token = NULL)
}
