% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuranker_domains.R
\name{accuranker_domains}
\alias{accuranker_domains}
\title{Get a list of all domains in the authorized Accuranker account}
\usage{
accuranker_domains(simple_token = NULL)
}
\arguments{
\item{simple_token}{Can be used to insert a custom API token instead of using oauth2.}
}
\description{
Get a list of all domains in the authorized Accuranker account
}
\examples{
accuranker_domains(simple_token = NULL)
}
