% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuranker_keywords.R
\name{accuranker_keywords}
\alias{accuranker_keywords}
\title{Get the keywords for a specific domain}
\usage{
accuranker_keywords(domain_name, silent = FALSE, simple_token = NULL)
}
\arguments{
\item{silent}{If TRUE, nothing will be printed to the console}

\item{simple_token}{Can be used to insert a custom API token instead of using oauth2.}

\item{domain_id}{A specific domain id}
}
\description{
Get the keywords for a specific domain
}
\examples{
accuranker_keywords(domain_name, silent = FALSE, simple_token = NULL)
}
